Refactor: move fetching to the preview step to improve the UX
